Renamed requester util.js to core.js, since it contains a lot of important functionalities
