+++ lib/requester/requester.js
@@ -1,5 +1,5 @@
 var _ = require('lodash'),
-    util = require('./util'),
+    core = require('./core'),
     Emitter = require('events'),
     inherits = require('inherits'),
     sdk = require('postman-collection'),
@@ -116,12 +116,12 @@ _.assign(Requester.prototype, /** @lends Requester.prototype */ {
         }
 
         cookieJar = self.options.cookieJar;
-        requestOptions = util.getRequestOptions(request, self.options);
+        requestOptions = core.getRequestOptions(request, self.options);
         startTime = Date.now();
         hostname = request.url.getHost();
 
         // check if host is on the `restrictedAddresses`
-        if (networkOptions.restrictedAddresses && util.isAddressRestricted(hostname, networkOptions)) {
+        if (networkOptions.restrictedAddresses && core.isAddressRestricted(hostname, networkOptions)) {
             return complete(new Error(ERROR_RESTRICTED_ADDRESS + hostname));
         }
 
@@ -143,14 +143,14 @@ _.assign(Requester.prototype, /** @lends Requester.prototype */ {
             responseString = ((resBody !== null && resBody !== undefined) && resBody.toString) ?
                 resBody.toString() : resBody;
             if (responseString === '[object ArrayBuffer]') {
-                responseString = util.arrayBufferToString(resBody);
+                responseString = core.arrayBufferToString(resBody);
             }
 
             // Calculate the time taken for us to get the response.
             responseTime = Date.now() - startTime;
 
             // This helps us to unify the information from XHR or Node calls.
-            responseJSON = util.jsonifyResponse(res, requestOptions, responseString);
+            responseJSON = core.jsonifyResponse(res, requestOptions, responseString);
 
             // Pull out cookies from the cookie jar, and make them chrome compatible.
             cookies = (cookieJar && _.isFunction(cookieJar.getCookies)) ?
